[Refactor] W7 피드백(페이징 부분 공통 처리)(#34)
