+++ src/main/java/com/example/umc10th/global/dto/CursorResponse.java
@@ -1,0 +1,9 @@
+package com.example.umc10th.global.dto;
+
+import java.util.List;
+
+public record CursorResponse<T>(
+        List<T> data,
+        Long cursor,
+        Boolean hasNext
+) {}
