+++ src/main/java/com/example/umc10th/global/dto/PageResponse.java
@@ -1,0 +1,9 @@
+package com.example.umc10th.global.dto;
+
+import java.util.List;
+
+public record PageResponse<T>(
+        List<T> data,
+        Integer pageNumber,
+        Integer pageSize
+) {}
